CHANGES:
Changed the backward functor operator in all the models to include the
current node value. This is especially useful in calibarting the trees.
This change is done in lattice_backward_traversals.h and
lattice_models.h. As a result of this I had to modify the examples and
testing modlues (lattice_model_t.h,lattice_algorithms_t.h).
Next, I started working on calibration of trees - introduced first
variation of the Calibratror object inside lattice_calibration.h.

TODO:

- Start working on Calibrator for IR asset class